Add canvas functionality to canvas component
